refactor: Complete Phase 2 OkHttp refactoring with improved role separation
## Phase 2 Implementation Summary

### Package Migration & Architecture
- Move OpenMockerInterceptor from `/interceptor/` → `/client/okhttp/`
- Integrate with Phase 1 abstraction layer (MockingEngine + OkHttpAdapter)
- Remove empty `/interceptor/` directory

### Role Separation Improvements
- Add `getMockData()` method to MockingEngine for clean mock data access
- Add `createMockResponse()` method to MockingEngine for consistent response creation
- Eliminate mixed usage of MockingEngine and CacheRepo in interceptor
- Achieve complete abstraction through MockingEngine interface

### Key Changes
- **OpenMockerInterceptor**: Now uses only MockingEngine methods, no direct CacheRepo access
- **MockingEngine**: Enhanced with `getMockData()` and `createMockResponse()` for role clarity
- **Import Updates**: Updated OpenMocker.kt to reference new package location
- **Delay Handling**: Optimized for OkHttp synchronous context with `runBlocking { delay() }`

### Backward Compatibility
- ✅ All existing unit tests pass (31 Phase 1 + library tests)
- ✅ Public API unchanged: `OpenMocker.getInterceptor()` works identically
- ✅ Demo app builds and integrates successfully
- ✅ Same behavior: caching, mocking, delays, response construction

### Technical Benefits
- **Code Reuse**: ~90% of mocking logic now shared through abstraction
- **Clean Architecture**: Clear separation between core logic and HTTP client specifics
- **Ktor Ready**: Architecture prepared for Phase 3 Ktor implementation
- **Maintainability**: Single responsibility principle with MockingEngine handling all mock operations

### Ktor Preparation
The improved role separation enables optimal Ktor support:
```kotlin
// OkHttp (synchronous)
runBlocking { delay(mockData.duration) }

// Ktor (asynchronous)
delay(mockData.duration) // Natural suspend function usage
```
